requirements.txt
loguru~=0.7.3
pyspark~=3.5.6
pendulum~=3.1.0
python-dotenv~=1.1.1